+++ src/backend.ts
@@ -223,7 +223,7 @@ export function GenStatistics(): Promise<LibraryStatistics> {
     return invoke("gen_statistics")
 }
 
-export function RecoverItems(params: { items: ItemId[] }): Promise<void> {
+export function RecoverItems(params: { items: ItemId[], parentOverride?: string }): Promise<void> {
     return invoke("recover_items", params)
 }